+++ packages/acp-bridge/src/index.ts
@@ -15,3 +15,4 @@ export * from './bridgeTypes.js';
 export * from './bridgeOptions.js';
 export * from './spawnChannel.js';
 export * from './bridgeClient.js';
+export * from './bridge.js';
